+++ api/prisma/migrations/20260801234810_add_location_status_expiry/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE `Location` ADD COLUMN `revertStatus` VARCHAR(15) NULL,
+    ADD COLUMN `statusExpiry` DATETIME(3) NULL;